# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "nadeemshaik1171998@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/createLabel": {
            "post": {
                "summary": "Create Label",
                "description": " Create Label ",
                "operationId": "c27db53decb5a6bcfdb5655c217ea42f",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "labelname"
                                ],
                                "properties": {
                                    "labelname": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label added Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    },
                    "401": {
                        "description": "Label Name already exists"
                    },
                    "405": {
                        "description": "Could not add label"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/displayLabelById": {
            "get": {
                "summary": "Display Label",
                "description": " Display Label ",
                "operationId": "a08f9443a445c7d17c0bd7b2b801b43f",
                "requestBody": {},
                "responses": {
                    "404": {
                        "description": "Label not Found"
                    },
                    "200": {
                        "description": "All Labels are Fetched Successfully"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/updateLabelById": {
            "post": {
                "summary": "Update Label",
                "description": " Update label ",
                "operationId": "3b0bb8f979dc9df0827cedfb893e5794",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "labelname"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "labelname": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Label updated Sucessfully"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    },
                    "404": {
                        "description": "Label not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/deleteLabelById": {
            "post": {
                "summary": "Delete Label",
                "description": " Delete label ",
                "operationId": "ac85f27dfdb7ba2b04dff974918c78af",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label deleted Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    },
                    "403": {
                        "description": "label not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/addLabelByNoteId": {
            "post": {
                "summary": "Add Label By Note Id",
                "description": " Add Label By Note Id ",
                "operationId": "a637e7ef841137bd7a3532f5703600e6",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "label_id",
                                    "note_id"
                                ],
                                "properties": {
                                    "label_id": {
                                        "type": "integer"
                                    },
                                    "note_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label notes added Successfully"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    },
                    "409": {
                        "description": "Note Already have a label"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/deleteNoteLabel": {
            "post": {
                "summary": "delete note label",
                "description": "delete note label",
                "operationId": "83506308dfb31c94ac89df41a52cabce",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "label_id",
                                    "note_id"
                                ],
                                "properties": {
                                    "label_id": {
                                        "description": "Request",
                                        "type": "integer"
                                    },
                                    "note_id": {
                                        "description": "Request",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label successfully deleted"
                    },
                    "404": {
                        "description": "Note not found with this label"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/displayNoteLabel": {
            "get": {
                "summary": "Display Label note",
                "description": " Display LabelNote ",
                "operationId": "e986299fe7196ca36e72664266d74587",
                "requestBody": {},
                "responses": {
                    "404": {
                        "description": "Label not Found"
                    },
                    "200": {
                        "description": "All Labels are Fetched Successfully"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}